1. Added configuration for FreeBSD freebsd-x86/config-host.h 2. added casting to uint32_t in skin_argb.h to avoid error during    compilation (incompatible type assignment) 3. added <sys/params.h> include. This allows correct definition of    namespaces and makes u_int32_t like types visible in <netinet/tcp.h> 4. Changed Makefile.android in order to build with OSS on FreeBSD    without prebuilt version of qemu_audio, which is now absent.

+++ skins/skin_argb.h
@@ -23,7 +23,7 @@ typedef  mmx_t  argb_t;
 static inline mmx_t
 mmx_load8888( unsigned  value, mmx_t  zero )
 {
-    return _mm_unpacklo_pi8( _mm_cvtsi32_si64 (value), zero);
+    return _mm_unpacklo_pi8( _mm_cvtsi32_si64 ((uint32_t)value), zero);
 }
 
 static inline unsigned
